// src/shared/types/IProductManual.ts
export interface IProductManual {
  id:          string;
  brand:       string;
  category:    null;
  code:        null;
  color:       null;
  currency:    string;
  description: string;
  editable:    number;
  image:       string;
  price:       string;
  stock:       number;
  title:       string;
  tableName:   string;
}